Refactor e2e tests using helpers
