+++ packages/zksync-dai-bridge/zksync-helpers/deploy.ts
@@ -1,0 +1,36 @@
+import { Deployer } from '@matterlabs/hardhat-zksync-deploy'
+import { Contract, Wallet } from 'ethers'
+import fs from 'fs'
+import { ethers } from 'hardhat'
+import * as hre from 'hardhat'
+import * as zk from 'zksync-web3'
+
+export async function deployL2Contract<T extends zk.Contract>(
+  l2Signer: zk.Wallet,
+  contractName: string,
+  args: Array<any> = [],
+): Promise<T> {
+  const deployer = new Deployer(hre, l2Signer)
+  const artifact = await deployer.loadArtifact(contractName)
+  const contract = await deployer.deploy(artifact, args)
+  console.log(`${contractName} was deployed on L2 to ${contract.address}`)
+  return contract as T
+}
+
+export async function deployL1Contract<T extends Contract>(
+  l1Signer: Wallet,
+  contractName: string,
+  args: Array<any> = [],
+): Promise<T> {
+  const jsonFilePath = `./artifacts/contracts/l1/${contractName}.sol/${contractName}.json`
+  if (!fs.existsSync(jsonFilePath)) {
+    throw new Error(`${jsonFilePath}  doesnt exist!`)
+  }
+  const json = JSON.parse(fs.readFileSync(jsonFilePath, 'utf-8'))
+  const factory = await ethers.getContractFactory(json.abi, json.bytecode)
+  const contractFactory = new ethers.ContractFactory(factory.interface, factory.bytecode, l1Signer)
+  const contractDeployed = await contractFactory.deploy(...args)
+  const contract = await contractDeployed.deployed()
+  console.log(`${contractName} was deployed on L1 to ${contract.address}`)
+  return contract as T
+}
