+++ packages/zksync-dai-bridge/zksync-helpers/index.ts
@@ -1,0 +1,4 @@
+export * from './bridge'
+export * from './deploy'
+export * from './messaging'
+export * from './mocks'
